backend/requirements.txt
amqp==2.2.2
asn1crypto==0.24.0
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.6
coreschema==0.0.4
cryptography==2.1.4
defusedxml==0.5.0
Django==2.0.2
django-cors-headers==2.1.0
django-extensions==1.9.9
django-filter==1.1.0
django-rest-swagger==2.1.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
drf-yasg==1.4.0
ecdsa==0.13
fix-yahoo-finance==0.0.21
future==0.16.0
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
more-itertools==4.1.0
multitasking==0.0.7
ndg-httpsclient==0.4.4
numpy==1.14.2
oauthlib==2.0.6
openapi-codec==1.3.2
pandas==0.22.0
pandas-datareader==0.6.0
pkg-resources==0.0.0
psycopg2==2.7.3.2
pyasn1==0.4.2
pycparser==2.18
pycryptodome==3.4.11
PyJWT==1.5.3
pyOpenSSL==17.5.0
python-dateutil==2.7.0
python-dotenv==0.7.1
python-jose==2.0.2
python3-openid==3.1.0
pytz==2017.3
Quandl==3.3.0
redis==2.10.6
requests==2.18.4
requests-file==1.4.3
requests-ftp==0.3.1
requests-oauthlib==0.8.0
ruamel.yaml==0.15.35
simplejson==3.13.2
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wrapt==1.10.11
